server/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
